fix: respect origin ReadV limits when forwarding proxy requests
Before the change all read requests (i.e. Read and ReadV) were forwarded
to the origin as ReadV by the proxy. That can cause errors if the length
of a Read request exceeds ReadV max_ior limit. To avoid the issue let's
forward Read requests as Read, not as ReadV.